# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PickSmart Academy + NOVA API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AssignmentStatus =
  (typeof AssignmentStatus)[keyof typeof AssignmentStatus];

export const AssignmentStatus = {
  pending: "pending",
  active: "active",
  paused: "paused",
  completed: "completed",
  archived: "archived",
} as const;

export type AssignmentVoiceMode =
  (typeof AssignmentVoiceMode)[keyof typeof AssignmentVoiceMode];

export const AssignmentVoiceMode = {
  training: "training",
  production: "production",
  ultra_fast: "ultra_fast",
} as const;

export interface Assignment {
  id: string;
  assignmentNumber: number;
  title: string;
  selectorUserId?: string;
  trainerUserId?: string;
  startAisle: number;
  endAisle: number;
  totalCases: number;
  totalCube: number;
  totalPallets: number;
  estimatedGold?: number;
  goalTimeMinutes?: number;
  goalTimeSeconds?: number;
  doorNumber: number;
  printerNumber: number;
  alphaLabelNumber: number;
  bravoLabelNumber: number;
  status: AssignmentStatus;
  voiceMode: AssignmentVoiceMode;
  startedAt?: string | null;
  completedAt?: string | null;
  totalDurationSeconds?: number | null;
  performancePercent?: number | null;
  percentComplete: number;
}

export interface CreateAssignmentBody {
  assignmentNumber: number;
  title: string;
  selectorUserId?: string;
  trainerUserId?: string;
  startAisle: number;
  endAisle: number;
  totalCases: number;
  totalCube: number;
  totalPallets: number;
  doorNumber: number;
  goalTimeMinutes?: number;
  goalTimeSeconds?: number;
  voiceMode?: string;
}

export interface UpdateAssignmentBody {
  status?: string;
  voiceMode?: string;
  selectorUserId?: string;
  percentComplete?: number;
  performancePercent?: number;
  startedAt?: string;
  completedAt?: string;
  totalDurationSeconds?: number;
}

export type AssignmentStopStatus =
  (typeof AssignmentStopStatus)[keyof typeof AssignmentStopStatus];

export const AssignmentStopStatus = {
  pending: "pending",
  arrived: "arrived",
  verified: "verified",
  picked: "picked",
} as const;

export interface AssignmentStop {
  id: string;
  assignmentId: string;
  stopOrder: number;
  aisle: number;
  slot: number;
  level?: string;
  checkCode: string;
  qty: number;
  status: AssignmentStopStatus;
  arrivedAt?: string | null;
  verifiedAt?: string | null;
  pickedAt?: string | null;
  dwellSeconds?: number | null;
}

export interface UpdateStopBody {
  status?: string;
  arrivedAt?: string;
  verifiedAt?: string;
  pickedAt?: string;
  dwellSeconds?: number;
}

export interface SystemDefaults {
  id: string;
  printerNumber: number;
  alphaLabelNumber: number;
  bravoLabelNumber: number;
}

export interface DoorCode {
  id: string;
  doorNumber: number;
  stagingCode: string;
  isActive: boolean;
}

export interface SlotEntry {
  id: string;
  aisle: number;
  slot: number;
  level?: string;
  checkCode: string;
  label: string;
  isActive: boolean;
}

export type TrainingModuleDifficulty =
  (typeof TrainingModuleDifficulty)[keyof typeof TrainingModuleDifficulty];

export const TrainingModuleDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface Lesson {
  id: string;
  title: string;
  videoUrl?: string | null;
  duration: number;
  description: string;
}

export interface TrainingModule {
  id: string;
  title: string;
  slug: string;
  description: string;
  difficulty: TrainingModuleDifficulty;
  duration: number;
  category: string;
  image: string;
  lessons: Lesson[];
  commonMistakes: string[];
  objectives: string[];
}

export interface LeaderboardEntry {
  userId: string;
  name: string;
  role: string;
  totalAssignments: number;
  avgPerformance: number;
  totalCases: number;
  rank: number;
}

export interface AssignmentSummary {
  total: number;
  active: number;
  pending: number;
  completed: number;
  avgPerformance: number;
  totalCasesToday: number;
}

export type ListSlotsParams = {
  search?: string;
};