# lib/api-zod/src/generated/types/assignment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PickSmart Academy + NOVA API
 * OpenAPI spec version: 0.1.0
 */
import type { AssignmentStatus } from "./assignmentStatus";
import type { AssignmentVoiceMode } from "./assignmentVoiceMode";

export interface Assignment {
  id: string;
  assignmentNumber: number;
  title: string;
  selectorUserId?: string;
  trainerUserId?: string;
  startAisle: number;
  endAisle: number;
  totalCases: number;
  totalCube: number;
  totalPallets: number;
  estimatedGold?: number;
  goalTimeMinutes?: number;
  goalTimeSeconds?: number;
  doorNumber: number;
  printerNumber: number;
  alphaLabelNumber: number;
  bravoLabelNumber: number;
  status: AssignmentStatus;
  voiceMode: AssignmentVoiceMode;
  startedAt?: string | null;
  completedAt?: string | null;
  totalDurationSeconds?: number | null;
  performancePercent?: number | null;
  percentComplete: number;
}