# lib/api-zod/src/generated/types/assignmentSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PickSmart Academy + NOVA API
 * OpenAPI spec version: 0.1.0
 */

export interface AssignmentSummary {
  total: number;
  active: number;
  pending: number;
  completed: number;
  avgPerformance: number;
  totalCasesToday: number;
}