# lib/api-zod/src/generated/types/createAssignmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PickSmart Academy + NOVA API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAssignmentBody {
  assignmentNumber: number;
  title: string;
  selectorUserId?: string;
  trainerUserId?: string;
  startAisle: number;
  endAisle: number;
  totalCases: number;
  totalCube: number;
  totalPallets: number;
  doorNumber: number;
  goalTimeMinutes?: number;
  goalTimeSeconds?: number;
  voiceMode?: string;
}