# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PickSmart Academy + NOVA API
 * OpenAPI spec version: 0.1.0
 */

export interface Lesson {
  id: string;
  title: string;
  videoUrl?: string | null;
  duration: number;
  description: string;
}