# lib/api-zod/src/generated/types/slotEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PickSmart Academy + NOVA API
 * OpenAPI spec version: 0.1.0
 */

export interface SlotEntry {
  id: string;
  aisle: number;
  slot: number;
  level?: string;
  checkCode: string;
  label: string;
  isActive: boolean;
}