# lib/api-zod/src/generated/types/systemDefaults.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PickSmart Academy + NOVA API
 * OpenAPI spec version: 0.1.0
 */

export interface SystemDefaults {
  id: string;
  printerNumber: number;
  alphaLabelNumber: number;
  bravoLabelNumber: number;
}