# lib/api-zod/src/generated/types/trainingModule.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PickSmart Academy + NOVA API
 * OpenAPI spec version: 0.1.0
 */
import type { Lesson } from "./lesson";
import type { TrainingModuleDifficulty } from "./trainingModuleDifficulty";

export interface TrainingModule {
  id: string;
  title: string;
  slug: string;
  description: string;
  difficulty: TrainingModuleDifficulty;
  duration: number;
  category: string;
  image: string;
  lessons: Lesson[];
  commonMistakes: string[];
  objectives: string[];
}