# lib/api-zod/src/generated/types/updateAssignmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PickSmart Academy + NOVA API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAssignmentBody {
  status?: string;
  voiceMode?: string;
  selectorUserId?: string;
  percentComplete?: number;
  performancePercent?: number;
  startedAt?: string;
  completedAt?: string;
  totalDurationSeconds?: number;
}