# lib/api-zod/src/generated/types/updateStopBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PickSmart Academy + NOVA API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateStopBody {
  status?: string;
  arrivedAt?: string;
  verifiedAt?: string;
  pickedAt?: string;
  dwellSeconds?: number;
}